Admin refactored to Dashboard.
