+++ frontend/src/assets/theme/components/dashboard-footer.js
