+++ frontend/src/assets/theme/components/dashboard-navbar.js
