+++ frontend/src/index.js
@@ -11,7 +11,7 @@ import "assets/plugins/nucleo/css/nucleo.css";
 import "@fortawesome/fontawesome-free/css/all.min.css";
 import "assets/scss/argon-dashboard-react.scss";
 
-import AdminLayout from "layouts/Admin.js";
+import AdminLayout from "layouts/Dashboard.js";
 import AuthLayout from "layouts/Auth.js";
 
 ReactDOM.render(
@@ -20,9 +20,9 @@ ReactDOM.render(
     <CssBaseline />
     <BrowserRouter>
       <Switch>
-        <Route path="/admin" render={(props) => <AdminLayout {...props} />} />
+        <Route path="/dashboard" render={(props) => <AdminLayout {...props} />} />
         <Route path="/auth" render={(props) => <AuthLayout {...props} />} />
-        <Redirect from="/" to="/admin/index" />
+        <Redirect from="/" to="/dashboard/index" />
       </Switch>
     </BrowserRouter>
   </ThemeProvider>,
